+++ src/cleanlab_tlm/utils/vpc/chat_completions.py
@@ -11,18 +11,18 @@ If you are not using VPC, use the `cleanlab_tlm.utils.chat_completions` module i
 from __future__ import annotations
 
 import os
-from typing import TYPE_CHECKING, Any, Optional
+from typing import TYPE_CHECKING, Any, Optional, Union, cast
 
 import requests
 
 from cleanlab_tlm.internal.base import BaseTLM
 from cleanlab_tlm.internal.constants import _VALID_TLM_QUALITY_PRESETS_CHAT_COMPLETIONS
+from cleanlab_tlm.tlm import TLMScore
+from cleanlab_tlm.utils.per_field_score_utils import _get_untrustworthy_fields
 from cleanlab_tlm.utils.vpc.tlm import VPCTLMOptions
 
 if TYPE_CHECKING:
     from openai.types.chat import ChatCompletion
-
-    from cleanlab_tlm.internal.types import JSONDict
 
 
 class TLMChatCompletion(BaseTLM):
@@ -74,7 +74,7 @@ class TLMChatCompletion(BaseTLM):
         *,
         response: ChatCompletion,
         **openai_kwargs: Any,
-    ) -> JSONDict:
+    ) -> TLMScore:
         """Score the trustworthiness of an OpenAI ChatCompletion response.
 
         Args:
@@ -84,11 +84,21 @@ class TLMChatCompletion(BaseTLM):
         Returns:
             TLMScore: A dict containing the trustworthiness score and optional logs
         """
+        try:
+            from openai.lib._parsing._completions import type_to_response_format_param
+        except ImportError as e:
+            raise ImportError(
+                f"OpenAI is required to use the {self.__class__.__name__} class. Please install it with `pip install openai`."
+            ) from e
+
         if (base_url := os.environ.get("BASE_URL")) is None:
             raise ValueError("BASE_URL is not set. Please set it in the environment variables.")
 
         # replace the model used for scoring with the specified model in options
         openai_kwargs["model"] = self._options["model"]
+
+        if "response_format" in openai_kwargs:
+            openai_kwargs["response_format"] = type_to_response_format_param(openai_kwargs["response_format"])
 
         res = requests.post(
             f"{base_url}/chat/score",
@@ -112,7 +122,49 @@ class TLMChatCompletion(BaseTLM):
 
         res_json = res.json()
         tlm_result = {"trustworthiness_score": res_json["tlm_metadata"]["trustworthiness_score"]}
-        if explanation := res_json["tlm_metadata"].get("log", {}).get("explanation"):
-            tlm_result["log"] = {"explanation": explanation}
 
-        return tlm_result
+        if self._return_log:
+            log = {}
+
+            log_options = cast(list[str], self._options.get("log", []))
+            if "explanation" in log_options:
+                explanation = res_json["tlm_metadata"].get("log", {}).get("explanation")
+                log["explanation"] = explanation
+
+            if "per_field_score" in log_options:
+                per_field_score = res_json["tlm_metadata"].get("log", {}).get("per_field_score")
+                log["per_field_score"] = per_field_score
+
+            tlm_result["log"] = log
+
+        return cast(TLMScore, tlm_result)
+
+    def get_untrustworthy_fields(
+        self,
+        *,
+        response: Optional[ChatCompletion] = None,
+        tlm_result: Union[TLMScore, ChatCompletion],
+        threshold: float = 0.8,
+        display_details: bool = True,
+    ) -> list[str]:
+        """Gets the fields of a structured output response that are considered untrustworthy by TLM.
+        Only works for responses that are valid JSON objects (uses `response_format` to specify the output format).
+        Prints detailed information about the untrustworthy fields if `display_details` is True.
+
+        Args:
+            response (ChatCompletion): The OpenAI ChatCompletion response object to evaluate
+            tlm_result (TLMScore | ChatCompletion): The result object from a previous TLM call
+            threshold (float): The threshold for considering a field untrustworthy
+            display_details (bool): Whether to display detailed information about the untrustworthy fields
+
+        Returns:
+            list[str]: The fields of the response that are considered untrustworthy by TLM
+        """
+
+        return _get_untrustworthy_fields(
+            response=response,
+            tlm_result=tlm_result,
+            threshold=threshold,
+            display_details=display_details,
+            class_name=self.__class__.__name__,
+        )